Delete utility on Products (Need to change db configuration)

+++ frontend/src/app/components/storepage/storepage.component.ts
@@ -82,7 +82,7 @@ export class StorepageComponent implements OnInit {
     
         dialogRef.afterClosed().subscribe(result => {
             console.log('Dialog Closed');
-            if (result.event === 'Delete') {
+            if (result !== undefined && result.event === 'Delete') {
                 this.delete();
             }
         });
